refactor: separate logic (callbacks, js callbacks, dynamic callbacks, scripts) to different files
